moved event subscription to its own file
